made clustering class
