+++ clustering/train_clusters.py
@@ -4,23 +4,23 @@ import pickle
 import numpy as np
 import h5py
 
-from dendrogram import cluster_groups, id_elms
-from elm_prediction import package_dir
+from clustering import Clustering
 from elm_prediction.train import train_loop
 
 
 if __name__ == '__main__':
 
-    with open(Path(__file__).parent / 'run_dir_classification/elm_predictions.pkl', 'r+b') as f:
-        elm_predictions = pickle.load(f)
+    run_dir = Path(__file__).parent / 'run_dir_classification'
+    dataset_dir = Path(__file__).parent / 'clustering_datasets'
+    clusters = Clustering(run_dir, dataset_dir)
 
-    ids_standard = id_elms(elm_predictions, scale='standard')
-    cluster = cluster_groups(elm_predictions, thresh=53, ids=ids_standard)
+    ids_standard = clusters.id_elms()
+    cluster = clusters.cluster_groups(thresh=53, ids=ids_standard)
     l_cluster = cluster[np.argmax([len(i) for i in cluster])]
 
     try:
         h = h5py.File('clustered_datasets/labeled_elm_events_cluster_1.hdf5', 'w-')
-        for elm, vals in elm_predictions.items():
+        for elm, vals in clusters.elm_predictions.items():
             if elm not in l_cluster:
                 continue
             signals = np.transpose(vals['signals'], (1, 2, 0)).reshape(64, -1)
@@ -45,4 +45,4 @@ if __name__ == '__main__':
             'output_dir': Path(__file__).parent / 'run_dir_classification_clustered',
             }
 
-    train_loop(args, select_elms=l_cluster)
+    train_loop(args)
